another fix after rubocop
