+++ 2016/PavelZhachko/1/homework_1.rb
@@ -1,7 +1,7 @@
 #!/usr/bin/env ruby
-require_relative "RPN_calculator"
+require_relative 'RPNCalculator'
 
 calc = RPNCalculator.new
-p "Input rpn ex - reverse polish notation example"
+p 'Input rpn ex - reverse polish notation example'
 a = gets.chomp
 calc.evaluate(a)
